% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supCPM.R
\name{supCPM}
\alias{supCPM}
\title{Run Supervised Capacity Preserved Mapping (supCPM)}
\usage{
supCPM(
  data,
  label,
  ratio = 0.7,
  no_dims = 2,
  compel_force = 1,
  dist = "euclidean",
  degree = 2,
  k = 7,
  niter1 = 500,
  niter2 = 700,
  seed = 40,
  factor = 1.3,
  verbose = T,
  init = T,
  epsilon = 1,
  lr = 500,
  intermediate = F
)
}
\arguments{
\item{data}{matrix; Input data matrix with rows representing cells and columns representing genes.}

\item{label}{vector; The label vector indicating which cluster that each cell belongs to.}

\item{ratio}{numeric between 0 and 1; This is a trade-off between the KL-divergence and trace ratio term. Default is 0.7.}

\item{no_dims}{integer; The dimensionality of the embedding space (default is 2).}

\item{compel_force}{0 or 1; Compel_force equals to 0 if user wants to pull clusters a bit apart, and equals to 1 if user wants the best preservation of the geometry (default is 0).}

\item{dist}{'euclidean' or 'geodesic'; Choice of the distance used in the high dimensions (default is 'euclidean').}

\item{degree}{numeric; The degree of freedom of the t-distribution in the high dimensions (default 2).}

\item{k}{integer; This controls the size of nearest neighbors when calculating the geodesic distance (default is 7).}

\item{niter1}{integer; Then iteration number of the first phase (default is 500).}

\item{niter2}{integer; The iteration of the second phase (default is 700).}

\item{seed}{integer; Random seed for supCPM. (default is 40)}

\item{factor}{numeric; The factor multiples on the high-dimensional distance matrix between different clusters.}

\item{verbose}{logical; Whether the progress of the objective function should be printed (default is TRUE).}

\item{init}{logical; Whether to use the random initialization or not. If the init=T,
then the initialization will be the MDS of the adjusted capacity distance matrix. (default is FALSE)}

\item{epsilon}{numeric; The multiple of the default epsilon, which is used in converting distance to probability; Default value is 1.}

\item{lr}{numeric; The learning rate for optimization. Default value is 500.}

\item{intermediate}{logic; Whether to output coordinates of the embedding points before switching objective function or not. Default is False.}
}
\value{
If intermediate=False, returns an matrix containing supCPM coordinates in the embedding dimenisons. If intermediate=True, return a list with the final embedding and the intermediate embedding points.
}
\description{
Run supCPM visualization on the data matrix. supCPM is a clustering guided visualization method designed for single-cell RNA data.
supCPM requires both high dimensional data and label information as input. Labels could be obtained by clustering algorithm or experiments.
With the adjustment of parameters, supCPM could also output the result of CPM (Capacity Preserving Mapping).
}
\references{
Zhiqian Zhai, Yu L. Lei, Rongrong Wang, Yuying Xie, Supervised Capacity Preserving Mapping: A Clustering Guided Visualization Method for scRNAseq data,
bioRxiv 2021.06.18.448900; doi: https://doi.org/10.1101/2021.06.18.448900
}
