% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualMetric.R
\name{VisualMetric}
\alias{VisualMetric}
\title{Visualization Quality Metrics}
\usage{
VisualMetric(Y, X, label, KncNeighbor, KnnNeighbor)
}
\arguments{
\item{Y}{matrix; The coordinates of the embedding data.}

\item{X}{matrix; The coordinates of the high-dimensional data.}

\item{label}{vector; The label vector indicating which cluster that each cell belongs to.}

\item{KncNeighbor}{integer; The size of nearest neighbor for KNC metric}

\item{KnnNeighbor}{integer; The size of nearest neighbor for KNN metric.}
}
\value{
Returns a list contains the following elements:
\item{CorV}{Spearman correlation between within-cluster total variation in the high-dimensional space and those in the low.}
\item{KNC}{The fraction of k-nearest neighbors of cluster means in the high dimensional data which are still preserved as the k-nearest cluster means in the embedding space.}
\item{CPD}{Spearman correlation between pairwise distances in the high and low dimensions.}
\item{CSM}{Cluster Separation Measure.}
\item{KNN}{The fraction of k-nearest neighbor pairs in the high dimensional data that are still$k-nearest neighbor pairs in the embedding dimensions.}
}
\description{
Visualization Quality Metrics
}
